\name{CA19_9}
\docType{data}
\alias{CA19_9}
\title{Reproducibility Example Dataset from CLSI EP05-A3}
\description{
  This data set consists of the example data of a complete three-site, multisample reproducibility study as presented 
  in the CLSI EP5-A3 guideline. 
  It shows quantitative results of an automated immunometric assay measuring parameter CA19-9. 
  This dataset is described in Appendix B of this guideline consisting of 6 samples, each measured on one of three
  sites, at five days with five replicates per day.
}
\usage{CA19_9}
\format{data.frame with 450 rows and 4 variables.}
\references{
       CLSI EP05-A3 - Evaluation of Precision of Quantitative Measurement Procedures; Approved Guideline - Third Edition.
       \href{www.clsi.org}{www.clsi.org}
}
\keyword{datasets}