% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vca.R
\name{check4MKL}
\alias{check4MKL}
\title{Check for Availability of Intel's Math Kernel Library.}
\usage{
check4MKL()
}
\value{
variable 'MKL' in envir "msgEnv" will be set to TRUE/FALSE
}
\description{
Majority of the code is borrowed from the Microsoft R Open Rprofile.site file.
In case MKL can be detected this information will be stored in a separate envrionment, which
other function know about. If so, an optimized version of function \code{\link{getGB}}
will be used which used ordinary matrix-objects instead of matrices defined by the
\code{Matrix}-package. This seems to accelerate computation time for large datasets
by up to factor 30.
}
\details{
This function is for internal use only and therefore not exported.
}
\author{
Authors of the Rprofile.site file in Microsoft R Open,
			Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}

