\name{dataEP05A2_2}
\docType{data}
\alias{dataEP05A2_2}
\title{Simulated Data of a CLSI EP05-A2 20/2/2 Experiment}
\description{
  This data set consists of simulated measurements for an experiment conducted to 
  evaluate the precision performance of measurement methods. On 20 days two separate
  runs with two replicates of the same sample are measured. Thus, factor 'day'
  is the top-level random factor (variance component), factor 'run' is nested within
  'day'.
}
\usage{data(dataEP05A2_2)}
\format{data.frame with 80 rows and 3 variables.}
\references{
       CLSI EP05-A2 - Evaluation of Precision Performance of Quantitative Measurement Methods.
       \href{https://clsi.org}{CLSI}
}
\keyword{datasets}