\name{dataEP05A2_example}
\docType{data}
\alias{dataEP05A2_example}
\title{Example Data from CLSI EP05-A2}
\description{
  This data set consists of the example data of a 20/2/2 experiment listed in the CLSI
  EP05-A2 guideline on page 25 for estimating intermediate precision, i.e. total variance for
  a fully-nested design with 3 variance components (day, run, error).\cr
  Note, that the results in the original EP05-A2 guideline were obtained using rounded intermediate results, whereas,
  package \code{VCA} uses full precision. Any differences between results listed in the CLSI EP05-A2 guideline
  and those generated by the package are due to error propagation in the working example presented in the CLSI guideline. 
  Here, full precision is used for all intermediate results.
}
\usage{dataEP05A2_example}
\format{data.frame with 80 rows and 3 variables.}
\references{
       CLSI EP05-A2 - Evaluation of Precision Performance of Quantitative Measurement Methods.
       \href{www.clsi.org}{www.clsi.org}
}
\keyword{datasets}

 

