\name{dataEP05A3_MS_3}
\docType{data}
\alias{dataEP05A3_MS_3}
\title{Simulated Data of a CLSI EP05-A3 3/5/5 Multi-Site Experiment}
\description{
  This data set consists of simulated measurements for an experiment to be conducted  
  for evaluation of the precision performance of measurement methods. On 3 sites, on 5 days
  5 replicates of the same sample are measured. Thus, factor 'site' is the top-level random 
  factor (variance component), factor 'day' is nested within 'site'.
}
\usage{data(dataEP05A3_MS_3)}
\format{data.frame with 75 rows and 3 variables.}
\references{
       Draft of CLSI EP05-A3 - Evaluation of Precision Performance of Quantitative Measurement Methods.
       \href{https://clsi.org}{CLSI}
}
\keyword{datasets}