\name{dataRS0003_1}
\docType{data}
\alias{dataRS0003_1}
\title{Simulated Repeated Measurements Data.}
\description{
  This data set consists of 21 measurements of the same sample, suitable
  to quantify the mesurement error on the same device.
}
\usage{data(dataRS0003_1)}
\format{data.frame with 21 rows and 1 variable.}
\keyword{datasets}