\name{dataRS0005_2}
\docType{data}
\alias{dataRS0005_2}
\title{Simulated Data of 5/3 Experiment.}
\description{
  This data set consists of 15 measurements of the same sample, 
  measured on 5 days with 3 measurements on each day. This small
  experiment is suitable to quantify between-day variability on the same device.
}
\usage{dataRS0005_2}
\format{data.frame with 15 rows and 2 variables.}
\keyword{datasets}