% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixef}
\alias{fixef}
\title{Generic Method for Extracting Fixed Effects from a Fitted Model}
\usage{
fixef(object, ...)
}
\arguments{
\item{object}{(object)}

\item{...}{additional parameters}
}
\description{
Generic Method for Extracting Fixed Effects from a Fitted Model
}
\seealso{
\code{\link{fixef.VCA}}
}

