% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vca.R
\name{getBasis}
\alias{getBasis}
\title{Get Orthogonal Basis of a Matrix.}
\usage{
getBasis(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{(object) two-dimensional, for which a Basis inverse
has to be computed}

\item{tol}{(numeric) tolerance value to be used in comparisons}
}
\value{
(object) A matrix whose columns form an orthogonal basis for the span of X
}
\description{
Calculate an orthogonal Basis for the span of the column vectors of the input matrix.
}
\author{
Florian Dufey \email{florian.dufey@contractors.roche.com}
}
