\name{Glucose}
\docType{data}
\alias{Glucose}
\title{Inermediate Precision Data from CLSI EP05-A3}
\description{
  This data set consists of the Glucose intermediate precision data in the CLSI
  EP05-A3 guideline, i.e. total variance for a fully-nested design with 3 variance components (day, run, error).\cr
  Note, that the results in the original EP05-A3 guideline were obtained using rounded intermediate results, whereas,
  package \code{VCA} uses full precision. Any differences between results listed in the CLSI EP05-A3 guideline
  and those generated by the package are due to error propagation in the working example presented in the CLSI guideline. 
  Here, full precision is used for all intermediate results.
}
\usage{data(Glucose)}
\format{data.frame with 80 rows and 3 variables.}
\references{
       CLSI EP05-A3 - Evaluation of Precision of Quantitative Measurement Procedures; Approved Guideline - Third Edition.
       \href{https://clsi.org}{CLSI}
}
\keyword{datasets}

 

