\name{HugeData}
\docType{data}
\alias{HugeData}
\title{Huge Dataset with Three Variables}
\description{
	This dataset was added to have a very large dataset available for the 
	unit-test suite. It is an unbalanced dataset with three variables and 8070
	observations.
}
\usage{data(HugeData)}
\format{A data frame with 8070 observations on the following 3 variables.
\itemize{
\item y\cr\cr
	The response variable.\cr\cr
\item VC1\cr\cr
	Variable with 8 levels corresponding to top-level variance component.\cr\cr
\item VC2\cr\cr
	Variable with 3920 levels corresponding to 2nd-level variance component.\cr
}}
\keyword{datasets}