% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vca.R
\name{lmerG}
\alias{lmerG}
\title{Construct Variance-Covariance Matrix of Random Effects for Models Fitted by Function 'lmer'.}
\usage{
lmerG(obj, cov = FALSE)
}
\arguments{
\item{obj}{(object) inheriting from class 'lmerMod'}

\item{cov}{(logical) TRUE = in case of non-zero covariances a block diagonal matrix will be constructed,
                 FALSE = a diagonal matrix with all off-diagonal element being equal to zero will be contructed}
}
\value{
(Matrix) representing the variance-covariance structure of random effects G
}
\description{
This function either restricts the variance-covariance matrix of random effects G to be either
diagonal ('cov=FALSE') or to take any non-zero covariances into account (default, 'cov=TRUE').
}
\details{
This function is not intended to be called directly by users and therefore not exported!
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}

