% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vca.R
\name{load_if_installed}
\alias{load_if_installed}
\title{Load 'RevoUtilsMath'-package if available.}
\usage{
load_if_installed(package)
}
\arguments{
\item{package}{(character) package name to load, usually this will be package
'RevoUtilsMath' if available}
}
\description{
This function is taken from the Rprofile.site file of Microsoft R Open.
It was added to the package namespace to avoid a NOTE during the R CMD check
process stating that this function is not gobally defined.
}
\details{
Only change to the original version is a different bracketing scheme to match
the one used in the remaining source-code of the package.
}
\author{
Authors of the Rprofile.site file in Microsoft R Open.
}
