\name{LSMeans_Data}
\docType{data}
\alias{LSMeans_Data}
\title{Dataset for Unit-Testing of LS Means}
\description{
  This data set is used for unit-testing LS Means functionality. Reference 
  results were generated in SAS PROC MIXED and rounded to two decimals as
  covariance parameter estimates slightly differ.
}
\usage{data(LSMeans_Data)}
\format{data.frame with 200 rows and 6 variables.}
\keyword{datasets}

 

