\name{MLrepro}
\docType{data}
\alias{MLrepro}
\title{Multi-Lot Reproducibility Data.}
\description{
  This data set consists of 754 observations. There are 3 laboratories (Lab), 3 lots (Lot),
  21 days (Days) per lab-lot combination, and 2 runs per day. The response variable is Result.
  This dataset is used in examples and unit-tests (see subdir 'UnitTests' of the package-dir). 
}
\usage{MLrepro}
\format{data.frame with 754 rows and 5 variables.}
\keyword{datasets}