% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.frame.VCA}
\alias{model.frame.VCA}
\title{Extract the Model Frame from a 'VCA' Object}
\usage{
\method{model.frame}{VCA}(formula, ...)
}
\arguments{
\item{formula}{(VCA) object}

\item{...}{additional arguments}
}
\value{
(data.frame) with attribute 'terms'
}
\description{
Function returns the data-element of 'object' and
adds the terms-element as attribute.
}
\details{
It enables application of functions relying on the existence of 
this method, e.g. the functin 'glht' of the 'multcomp'
R-package.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}

