% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.matrix.VCA}
\alias{model.matrix.VCA}
\title{Model Matrix of a Fitted VCA-Object}
\usage{
\method{model.matrix}{VCA}(object, ...)
}
\arguments{
\item{object}{(VCA) object}

\item{...}{further arguments}
}
\description{
Function returns matrix \code{X} corresponding
to the design matrix of fixed effects of the fitted
model.
}

