\name{Orthodont}
\docType{data}
\alias{Orthodont}
\title{Orthodont dataset from R-package 'nlme'}
\description{
	The Orthodont data frame has 108 rows and 4 columns of the change in an orthdontic measurement over time for several young subjects. 
	
	This dataset was included to simplify its usage in automated unit-tests (see directory UnitTests) and examples. 
	
	Investigators at the University of North Carolina Dental School followed the growth of 27 children (16 males, 11 females) 
	from age 8 until age 14. Every two years they measured the distance between the pituitary and the pterygomaxillary fissure, 
	two points that are easily identified on x-ray exposures of the side of the head. 
}
\usage{Orthodont}
\format{data.frame with 80 rows and 3 variables.}
\references{
	Pinheiro, J. C. and Bates, D. M. (2000), Mixed-Effects Models in S and S-PLUS, Springer, New York. (Appendix A.17)		

	Potthoff, R. F. and Roy, S. N. (1964), A generalized multivariate analysis of variance model useful especially for growth curve problems, Biometrika, 51, 313-326. 
}
\keyword{datasets}
