\name{print.VCA}
\alias{print.VCA}
\title{Standard Printing Method for Objects of Class 'VCA'.}
\usage{
\method{print}{VCA}(x, digits = 6L, ...)
}
\arguments{
  \item{x}{(VCA) object of class 'VCA' as returned by
  function 'anovaVCA'.}

  \item{digits}{(integer) number of digits numeric values
  are rounded to before printing.}

  \item{...}{additional arguments to be passed to or from
  methods.}
}
\description{
Function prints 'VCA' objects as returned by function
\code{\link{anovaVCA}}.
}

