\name{ranef}
\alias{ranef}
\title{Generic Method for Extracting Random Effects from a Fitted Model.}
\usage{
ranef(object, ...)
}
\arguments{
  \item{object}{(object)}

  \item{...}{additional parameters}
}
\description{
Generic Method for Extracting Random Effects from a Fitted
Model.
}

