\name{realData}
\docType{data}
\alias{realData}
\title{Real-World Data}
\description{
  This dataset is meant to serve as real-world representative completing the collection
  of datasets coming with this package. There are 6 variables, one response variable ('y')
  corresponding to concentration values of the measurand, and 5 factor variables. Variable "calibration"
  corresponds to the day a (re-) calibration was performed, all other variables are more or less
  self-explaining.
}
\usage{data(realData)}
\format{data.frame with 2268 rows and 6 variables.}
\keyword{datasets}