\name{ReproData1}
\docType{data}
\alias{ReproData1}
\title{Multi-Site Data for Estimating Reproducibility Precision}
\description{
  This data set consists of real-world measurements of a multi-site study aiming at quantifying
  reproducibility precision. Unlike in the CLSI EP05-A3 guideline, there are two runs per day with
  three replicated measurements per run. 
}
\usage{data(ReproData1)}
\format{data.frame with 120 rows and 4 variables.}
\references{
       Approved Guideline CLSI EP05-A3 - Evaluation of Precision Performance of Quantitative Measurement Methods.
       \href{https://clsi.org}{CLSI}
}
\keyword{datasets}