% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleData}
\alias{scaleData}
\title{Scale Response Variable to Ensure Robust Numerical Calculations}
\usage{
scaleData(Data = NULL, resp = NULL)
}
\arguments{
\item{Data}{(data.frame) with the data to be fitted and the response to be scaled}

\item{resp}{(character) name of the (numeric) response variable}
}
\value{
(data.frame) with the response scaled according to the scaling-factor,
which is recorded in the attribute \code{scale} of the data set
}
\description{
Function determines scaling factor for transforming the mean of the response to a range
between 0.1 and 1, applies scaling of the response and binds the scaling factor to the 
data as attribute 'scale'.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmester@roche.com}
}
