\name{sleepstudy}
\docType{data}
\alias{sleepstudy}
\title{sleepstudy dataset from R-package 'lme4'}
\description{
	The average reaction time per day for subjects in a sleep deprivation study. 
	On day 0 the subjects had their normal amount of sleep. Starting that night 
	they were restricted to 3 hours of sleep per night. The observations represent 
	the average reaction time on a series of tests given each day to each subject. 
}
\usage{sleepstudy}
\format{A data frame with 180 observations on the following 3 variables.
\itemize{
\item Reaction\cr\cr
	Average reaction time (ms)\cr\cr
\item Days\cr\cr
	Number of days of sleep deprivation\cr\cr
\item Subject\cr\cr
	Subject number on which the observation was made.\cr
}}
\references{
	Gregory Belenky, Nancy J. Wesensten, David R. Thorne, Maria L. Thomas, 
	Helen C. Sing, Daniel P. Redmond, Michael B. Russo and Thomas J. Balkin (2003) 
	Patterns of performance degradation and restoration during sleep restriction and 
	subsequent recovery: a sleep dose-response study. Journal of Sleep Research 12, 1-12.
}
\keyword{datasets}