\name{VCAdata1}
\docType{data}
\alias{VCAdata1}
\title{Simulated Data for Variance Component Analysis.}
\description{
  This data set consists of 2520 observations. There are 3 lots (lot),
  10 samples, 21 days, 2 runs within day. This simulated dataset is used in 
  examples and unit-tests (see subdir 'UnitTests' of the package-dir).
}
\usage{data(VCAdata1)}
\format{data.frame with 2520 rows and 5 variables.}
\keyword{datasets}