\name{fixef}
\alias{fixef}
\title{Generic Method for Extracting Fixed Effects from a Fitted Model.}
\usage{
fixef(object, ...)
}
\arguments{
  \item{object}{(object)}

  \item{...}{additional parameters}
}
\description{
Generic Method for Extracting Fixed Effects from a Fitted
Model.
}

