% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vca.R
\name{ranef}
\alias{ranef}
\title{Generic Method for Extracting Random Effects from a Fitted Model.}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{(object)}

\item{...}{additional parameters}
}
\description{
Generic Method for Extracting Random Effects from a Fitted Model.
}
\seealso{
\code{\link{ranef.VCA}}
}

