% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vca.R
\name{Trace}
\alias{Trace}
\title{Compute the Trace of a Matrix.}
\usage{
Trace(x, quiet = FALSE)
}
\arguments{
\item{x}{(matrix, Matrix) object}

\item{quiet}{(logical) TRUE = will suppress any warning, which will be issued otherwise}
}
\value{
(numeric) value, the trace of the matrix
}
\description{
Function computes the sum of main-diagonal elements of a squared matrix.
}

